requirements.txt
fastapi
uvicorn[standard]
python-dotenv
jinja2
python-multipart
langchain
langchain-community
langchain-google-genai
faiss-cpu
google-generativeai
firebase-admin
reportlab
gTTS